# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'API Gestion Immobilière Documentation'
  description: "Documentation complète de l'API REST pour la gestion de biens immobiliers"
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: null
security:
  -
    default: []
paths:
...
  /api/properties:
    get:
      summary: 'List Properties'
      operationId: listProperties
      description: 'Obtenir la liste paginée des biens'
      parameters:
        -
          in: query
          name: city
          description: 'Filtrer par ville.'
          example: Alger
          required: false
          schema:
            type: string
            description: 'Filtrer par ville.'
            example: Alger
            nullable: false
        -
          in: query
          name: type
          description: 'Type de bien.'
          example: Villa
          required: false
          schema:
            type: string
            description: 'Type de bien.'
            example: Villa
            nullable: false
        -
          in: query
          name: min_price
          description: 'Prix minimum.'
          example: 20000000.0
          required: false
          schema:
            type: number
            description: 'Prix minimum.'
            example: 20000000.0
            nullable: false
        -
          in: query
          name: max_price
          description: 'Prix maximum.'
          example: 50000000.0
          required: false
          schema:
            type: number
            description: 'Prix maximum.'
            example: 50000000.0
            nullable: false
        -
          in: query
          name: status
          description: Statut.
          example: disponible
          required: false
          schema:
            type: string
            description: Statut.
            example: disponible
            nullable: false
        -
          in: query
          name: search
          description: Recherche.
          example: piscine
          required: false
          schema:
            type: string
            description: Recherche.
            example: piscine
            nullable: false
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"data\": [...]\n}"
      tags:
        - 'Biens Immobiliers'
      security: []
    post:
      summary: 'Create Property'
      operationId: createProperty
      description: 'Créer un nouveau bien (Agent/Admin)'
      parameters: []
      responses:
        201:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"success\": true,\n  \"data\": {...}\n}"
      tags:
        - 'Biens Immobiliers'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  description: 'Type (Villa, Appartement, etc).'
                  example: Villa
                  nullable: false
                rooms:
                  type: integer
                  description: 'Nombre de pièces.'
                  example: 4
                  nullable: true
                surface:
                  type: number
                  description: 'Surface en m².'
                  example: 200.0
                  nullable: false
                price:
                  type: number
                  description: 'Prix en DZD.'
                  example: 25000000.0
                  nullable: false
                city:
                  type: string
                  description: Ville.
                  example: Alger
                  nullable: false
                neighborhood:
                  type: string
                  description: Quartier.
                  example: Hydra
                  nullable: true
                description:
                  type: string
                  description: Description.
                  example: 'Belle villa'
                  nullable: false
                status:
                  type: string
                  description: Statut.
                  example: disponible
                  nullable: false
                published:
                  type: boolean
                  description: Publié.
                  example: true
                  nullable: false
              required:
                - type
                - surface
                - price
                - city
                - description
                - status
  '/api/properties/{id}':
    get:
      summary: 'Show Property'
      operationId: showProperty
      description: "Afficher le détail d'un bien"
      parameters: []
      responses:
        200:
          description: ''
          content:
            text/plain:
              schema:
                type: string
                example: "{\n  \"data\": {...}\n}"
      tags:
        - 'Biens Immobiliers'
      security: []
    put:
      summary: 'Update Property'
      operationId: updateProperty
      description: 'Modifier un bien (Propriétaire/Admin)'
      parameters: []
      responses: {  }
      tags:
        - 'Biens Immobiliers'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  description: ''
                  example: Duplex
                  nullable: false
                  enum:
                    - Appartement
                    - Villa
                    - Terrain
                    - Studio
                    - Duplex
                rooms:
                  type: integer
                  description: 'Must be at least 1.'
                  example: 16
                  nullable: true
                surface:
                  type: number
                  description: 'Must be at least 1.'
                  example: 22
                  nullable: false
                price:
                  type: number
                  description: 'Must be at least 0.'
                  example: 84
                  nullable: false
                city:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: z
                  nullable: false
                neighborhood:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: m
                  nullable: true
                description:
                  type: string
                  description: ''
                  example: 'Eius et animi quos velit et.'
                  nullable: false
                status:
                  type: string
                  description: ''
                  example: vendu
                  nullable: false
                  enum:
                    - disponible
                    - vendu
                    - location
                published:
                  type: boolean
                  description: ''
                  example: true
                  nullable: false
              required:
                - type
                - surface
                - price
                - city
                - description
                - status
    delete:
      summary: 'Delete Property'
      operationId: deleteProperty
      description: 'Supprimer un bien (soft delete)'
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  success: true
                  message: 'Bien supprimé'
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: 'Bien supprimé'
      tags:
        - 'Biens Immobiliers'
    parameters:
      -
        in: path
        name: id
        description: 'ID du bien.'
        example: 1
        required: true
        schema:
          type: integer
...